Moved reconcile code into a new service and recorded new cassettes
