+++ spec/support/vcr_cassettes/form1010_ezr/veteran_enrollment_system/associations/create_associations_success.yml
@@ -1,0 +1,108 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: "<VETERAN_ENROLLMENT_SYSTEM_BASE_URI>/ves-associate-gateway-svc/associations/person/1012829228V424035"
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept:
+      - application/json
+      Content-Type:
+      - application/json
+      User-Agent:
+      - Vets.gov Agent
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Thu, 05 Jun 2025 20:31:42 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Server:
+      - envoy
+      X-Frame-Options:
+      - SAMEORIGIN
+      Content-Security-Policy:
+      - 'default-src https: data: ''unsafe-inline'' ''unsafe-eval'''
+      X-Envoy-Upstream-Service-Time:
+      - '23'
+      Ssl-Env:
+      - 'On'
+      X-Xss-Protection:
+      - 1; mode=block
+      X-Content-Type-Options:
+      - nosniff
+      Server-Timing:
+      - dtSInfo;desc="0", dtRpid;desc="1292492122"
+      Strict-Transport-Security:
+      - max-age=16000000; includeSubDomains; preload;
+    body:
+      encoding: UTF-8
+      string: '{"data":{"associations":[]},"messages":[]}'
+  recorded_at: Thu, 05 Jun 2025 20:31:42 GMT
+- request:
+    method: put
+    uri: "<VETERAN_ENROLLMENT_SYSTEM_BASE_URI>/ves-associate-gateway-svc/associations/person/1012829228V424035"
+    body:
+      encoding: UTF-8
+      string: '{"associations":[{"address":{"city":"Ontario","country":"CAN","line1":"834
+        NE 36th St","provinceCode":"ON","postalCode":"21534"},"name":{"givenName":"FIRSTNOKB","middleName":"MIDDLENOKB","familyName":"LASTNOKB"},"alternatePhone":"2658351234","primaryPhone":"1239183546","relationType":"DAUGHTER","role":"OTHER_NEXT_OF_KIN","lastUpdateDate":"2025-06-05T20:31:42Z"},{"address":{"city":"Blacksburg","country":"USA","line1":"823
+        SE 2nd St","state":"VA","zipCode":"24060"},"name":{"givenName":"FIRSTECB","middleName":"MIDDLEECB","familyName":"LASTECB"},"alternatePhone":"9563740017","primaryPhone":"9564738265","relationType":"UNRELATED_FRIEND","role":"OTHER_EMERGENCY_CONTACT","lastUpdateDate":"2025-06-05T20:31:42Z"},{"address":{"city":"Dulles","country":"USA","line1":"476
+        Test Ave","state":"VA","zipCode":"20101","zipPlus4":"0101"},"name":{"givenName":"FIRSTNOKA","middleName":"MIDDLENOKA","familyName":"LASTNOKA","suffix":"JR."},"alternatePhone":"7254551234","primaryPhone":"1239131234","relationType":"SON","role":"PRIMARY_NEXT_OF_KIN","lastUpdateDate":"2025-06-05T20:31:42Z"},{"address":{"city":"Dulles","country":"USA","line1":"28
+        NW 78th St","state":"VA","zipCode":"20101","zipPlus4":"0101"},"name":{"givenName":"FIRSTECA","middleName":"MIDDLEECA","familyName":"LASTECA"},"alternatePhone":"2699352134","primaryPhone":"7452743546","relationType":"BROTHER","role":"EMERGENCY_CONTACT","lastUpdateDate":"2025-06-05T20:31:42Z"}]}'
+    headers:
+      Accept:
+      - application/json
+      Content-Type:
+      - application/json
+      User-Agent:
+      - Vets.gov Agent
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Thu, 05 Jun 2025 20:31:44 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Server:
+      - envoy
+      X-Frame-Options:
+      - SAMEORIGIN
+      Content-Security-Policy:
+      - 'default-src https: data: ''unsafe-inline'' ''unsafe-eval'''
+      X-Envoy-Upstream-Service-Time:
+      - '370'
+      Ssl-Env:
+      - 'On'
+      X-Xss-Protection:
+      - 1; mode=block
+      X-Content-Type-Options:
+      - nosniff
+      Server-Timing:
+      - dtSInfo;desc="0", dtRpid;desc="758072780"
+      Strict-Transport-Security:
+      - max-age=16000000; includeSubDomains; preload;
+    body:
+      encoding: UTF-8
+      string: '{"data":{"associations":[{"role":"OTHER_NEXT_OF_KIN","name":{"givenName":"FIRSTNOKB","middleName":"MIDDLENOKB","familyName":"LASTNOKB"},"status":"INSERTED"},{"role":"OTHER_EMERGENCY_CONTACT","name":{"givenName":"FIRSTECB","middleName":"MIDDLEECB","familyName":"LASTECB"},"status":"INSERTED"},{"role":"PRIMARY_NEXT_OF_KIN","name":{"givenName":"FIRSTNOKA","middleName":"MIDDLENOKA","familyName":"LASTNOKA","suffix":"JR."},"status":"INSERTED"},{"role":"EMERGENCY_CONTACT","name":{"givenName":"FIRSTECA","middleName":"MIDDLEECA","familyName":"LASTECA"},"status":"INSERTED"}]},"messages":[{"description":"The
+        request was successfully processed.","code":"completed_success"}]}'
+  recorded_at: Thu, 05 Jun 2025 20:31:44 GMT
+recorded_with: VCR 6.3.1
