+++ spec/support/vcr_cassettes/form1010_ezr/veteran_enrollment_system/associations/update_associations_partial_success.yml
@@ -1,0 +1,116 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: "<VETERAN_ENROLLMENT_SYSTEM_BASE_URI>/ves-associate-gateway-svc/associations/person/1012829228V424035"
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept:
+      - application/json
+      Content-Type:
+      - application/json
+      User-Agent:
+      - Vets.gov Agent
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Thu, 05 Jun 2025 21:23:41 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Server:
+      - envoy
+      X-Frame-Options:
+      - SAMEORIGIN
+      Content-Security-Policy:
+      - 'default-src https: data: ''unsafe-inline'' ''unsafe-eval'''
+      X-Envoy-Upstream-Service-Time:
+      - '22'
+      Ssl-Env:
+      - 'On'
+      X-Xss-Protection:
+      - 1; mode=block
+      X-Content-Type-Options:
+      - nosniff
+      Server-Timing:
+      - dtSInfo;desc="0", dtRpid;desc="-1283945747"
+      Strict-Transport-Security:
+      - max-age=16000000; includeSubDomains; preload;
+    body:
+      encoding: UTF-8
+      string: '{"data":{"associations":[{"role":"OTHER_NEXT_OF_KIN","relationType":"CHILDINLAW","name":{"givenName":"UPDATEDFIRSTNOKB","middleName":"UPDATEDMIDDLENOKB","familyName":"UPDATEDLASTNOKB"},"address":{"line1":"875
+        Updated Blvd","line2":"Unit 532","city":"Tampa","state":"FL","zipCode":"33726","zipPlus4":"3942","country":"USA"},"primaryPhone":"1238835546","alternatePhone":"2658350023","lastUpdateDate":"2025-06-05T16:21:07"},{"role":"OTHER_EMERGENCY_CONTACT","relationType":"GRANDCHILD","name":{"givenName":"UPDATEDFIRSTECB","middleName":"UPDATEDMIDDLEECB","familyName":"UPDATEDLASTECB"},"address":{"line1":"875
+        West Blvd","line2":"Apt 3","line3":"Unit 6","city":"Wichita","state":"KS","zipCode":"67203","zipPlus4":"1234","country":"USA"},"primaryPhone":"6437432434","alternatePhone":"9563001117","lastUpdateDate":"2025-06-05T16:21:07"},{"role":"PRIMARY_NEXT_OF_KIN","relationType":"NIECE_NEPHEW","name":{"givenName":"UPDATEDFIRSTNOKA","middleName":"UPDATEDMIDDLENOKA","familyName":"UPDATEDLASTNOKA","suffix":"JR."},"address":{"line1":"NE
+        54th St","line2":"Apt 7","line3":"Unit 1222","city":"guanajuato","provinceCode":"guanajuato","postalCode":"84754","country":"MEX"},"primaryPhone":"4449131234","alternatePhone":"6544551234","lastUpdateDate":"2025-06-05T16:21:07"},{"role":"EMERGENCY_CONTACT","relationType":"EXTENDED_FAMILY_MEMBER","name":{"givenName":"UPDATEDFIRSTECA","middleName":"UPDATEDMIDDLEECA","familyName":"UPDATEDLASTECA"},"address":{"line1":"28
+        Parker St","city":"Los Angeles","state":"CA","zipCode":"90038","zipPlus4":"1234","country":"USA"},"primaryPhone":"7563627422","alternatePhone":"1123321232","lastUpdateDate":"2025-06-05T16:21:07"}]},"messages":[]}'
+  recorded_at: Thu, 05 Jun 2025 21:23:41 GMT
+- request:
+    method: put
+    uri: "<VETERAN_ENROLLMENT_SYSTEM_BASE_URI>/ves-associate-gateway-svc/associations/person/1012829228V424035"
+    body:
+      encoding: UTF-8
+      string: '{"associations":[{"address":{"city":"guanajuato","country":"MEX","line1":"NE
+        54th St","line2":"Apt 7","line3":"Unit 1222","provinceCode":"guanajuato","postalCode":"84754"},"name":{"givenName":"UPDATEDFIRSTNOKA","middleName":"UPDATEDMIDDLENOKA","familyName":"UPDATEDLASTNOKA","suffix":"JR."},"alternatePhone":"6544551234","deleteIndicator":true,"primaryPhone":"4449131234","relationType":"NIECE_NEPHEW","role":"PRIMARY_NEXT_OF_KIN"},{"address":{"city":"Tampa","country":"USA","line1":"875
+        Updated Blvd","line2":"Unit 532","state":"FL","zipCode":"33726","zipPlus4":"3942"},"name":{"givenName":"UPDATEDFIRSTNOKB","middleName":"UPDATEDMIDDLENOKB","familyName":"UPDATEDLASTNOKB"},"alternatePhone":"2658350023","deleteIndicator":true,"primaryPhone":"1238835546","relationType":"CHILDINLAW","role":"OTHER_NEXT_OF_KIN"},{"address":{"city":"Los
+        Angeles","country":"USA","line1":"28 Parker St","state":"CA","zipCode":"90038","zipPlus4":"1234"},"name":{"givenName":"UPDATEDFIRSTECA","middleName":"UPDATEDMIDDLEECA","familyName":"UPDATEDLASTECA"},"alternatePhone":"1123321232","deleteIndicator":true,"primaryPhone":"7563627422","relationType":"EXTENDED_FAMILY_MEMBER","role":"EMERGENCY_CONTACT"},{"address":{"city":"Wichita","country":"USA","line1":"875
+        West Blvd","line2":"Apt 3","line3":"Unit 6","state":"KS","zipCode":"67203","zipPlus4":"1234"},"name":{"givenName":"UPDATEDFIRSTECB","middleName":"UPDATEDMIDDLEECB","familyName":"UPDATEDLASTECB"},"alternatePhone":"9563001117","deleteIndicator":true,"primaryPhone":"6437432434","relationType":"GRANDCHILD","role":"OTHER_EMERGENCY_CONTACT"}]}'
+    headers:
+      Accept:
+      - application/json
+      Content-Type:
+      - application/json
+      User-Agent:
+      - Vets.gov Agent
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Thu, 05 Jun 2025 21:23:49 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Server:
+      - envoy
+      X-Frame-Options:
+      - SAMEORIGIN
+      Content-Security-Policy:
+      - 'default-src https: data: ''unsafe-inline'' ''unsafe-eval'''
+      X-Envoy-Upstream-Service-Time:
+      - '197'
+      Ssl-Env:
+      - 'On'
+      X-Xss-Protection:
+      - 1; mode=block
+      X-Content-Type-Options:
+      - nosniff
+      Server-Timing:
+      - dtSInfo;desc="0", dtRpid;desc="-1494902027"
+      Strict-Transport-Security:
+      - max-age=16000000; includeSubDomains; preload;
+    body:
+      encoding: UTF-8
+      string: '{"data":{"associations":[{"role":"PRIMARY_NEXT_OF_KIN","name":{"givenName":"UPDATEDFIRSTNOKA","middleName":"UPDATEDMIDDLENOKA","familyName":"UPDATEDLASTNOKA","suffix":"JR."},"status":"DELETED"},{"role":"OTHER_NEXT_OF_KIN","name":{"givenName":"UPDATEDFIRSTNOKB","middleName":"UPDATEDMIDDLENOKB","familyName":"UPDATEDLASTNOKB"},"status":"NOT_DELETED_NO_MATCHING_ASSOCIATION","message":"The
+        association was not deleted for this record, as there are no matching association
+        on file."},{"role":"EMERGENCY_CONTACT","name":{"givenName":"UPDATEDFIRSTECA","middleName":"UPDATEDMIDDLEECA","familyName":"UPDATEDLASTECA"},"status":"DELETED"},{"role":"OTHER_EMERGENCY_CONTACT","name":{"givenName":"UPDATEDFIRSTECB","middleName":"UPDATEDMIDDLEECB","familyName":"UPDATEDLASTECB"},"status":"NOT_DELETED_NO_MATCHING_ASSOCIATION","message":"The
+        association was not deleted for this record, as there are no matching association
+        on file."}]},"messages":[{"description":"Not all of the association records
+        were updated","code":"completed_partial"}]}'
+  recorded_at: Thu, 05 Jun 2025 21:23:49 GMT
+recorded_with: VCR 6.3.1
