+++ spec/support/vcr_cassettes/form1010_ezr/veteran_enrollment_system/associations/update_associations_success.yml
@@ -1,0 +1,112 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: "<VETERAN_ENROLLMENT_SYSTEM_BASE_URI>/ves-associate-gateway-svc/associations/person/1012829228V424035"
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept:
+      - application/json
+      Content-Type:
+      - application/json
+      User-Agent:
+      - Vets.gov Agent
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Thu, 05 Jun 2025 20:59:59 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Server:
+      - envoy
+      X-Frame-Options:
+      - SAMEORIGIN
+      Content-Security-Policy:
+      - 'default-src https: data: ''unsafe-inline'' ''unsafe-eval'''
+      X-Envoy-Upstream-Service-Time:
+      - '36'
+      Ssl-Env:
+      - 'On'
+      X-Xss-Protection:
+      - 1; mode=block
+      X-Content-Type-Options:
+      - nosniff
+      Server-Timing:
+      - dtSInfo;desc="0", dtRpid;desc="1146187143"
+      Strict-Transport-Security:
+      - max-age=16000000; includeSubDomains; preload;
+    body:
+      encoding: UTF-8
+      string: '{"data":{"associations":[{"role":"OTHER_NEXT_OF_KIN","relationType":"CHILDINLAW","name":{"givenName":"UPDATEDFIRSTNOKB","middleName":"UPDATEDMIDDLENOKB","familyName":"UPDATEDLASTNOKB"},"address":{"line1":"845
+        Glendale Ave","line2":"Unit 43","city":"Clearwater","state":"FL","zipCode":"33754","zipPlus4":"8753","country":"USA"},"primaryPhone":"1238835546","alternatePhone":"2658350023","lastUpdateDate":"2025-06-05T15:51:53"},{"role":"OTHER_EMERGENCY_CONTACT","relationType":"GRANDCHILD","name":{"givenName":"UPDATEDFIRSTECB","middleName":"UPDATEDMIDDLEECB","familyName":"UPDATEDLASTECB"},"address":{"line1":"875
+        West Blvd","line2":"Apt 3","line3":"Unit 6","city":"Wichita","state":"KS","zipCode":"67203","zipPlus4":"1234","country":"USA"},"primaryPhone":"9942738265","alternatePhone":"9563001117","lastUpdateDate":"2025-06-05T15:51:53"},{"role":"PRIMARY_NEXT_OF_KIN","relationType":"NIECE_NEPHEW","name":{"givenName":"UPDATEDFIRSTNOKA","middleName":"UPDATEDMIDDLENOKA","familyName":"UPDATEDLASTNOKA","suffix":"JR."},"address":{"line1":"SW
+        54th St","line2":"Apt 1","line3":"Unit 4","city":"chihuahua","provinceCode":"chihuahua","postalCode":"54345","country":"MEX"},"primaryPhone":"4449131234","alternatePhone":"6544551234","lastUpdateDate":"2025-06-05T15:51:53"},{"role":"EMERGENCY_CONTACT","relationType":"EXTENDED_FAMILY_MEMBER","name":{"givenName":"UPDATEDFIRSTECA","middleName":"UPDATEDMIDDLEECA","familyName":"UPDATEDLASTECA"},"address":{"line1":"28
+        Parker St","city":"Los Angeles","state":"CA","zipCode":"90038","zipPlus4":"1234","country":"USA"},"primaryPhone":"3322743546","alternatePhone":"2694437134","lastUpdateDate":"2025-06-05T15:51:53"}]},"messages":[]}'
+  recorded_at: Thu, 05 Jun 2025 20:59:59 GMT
+- request:
+    method: put
+    uri: "<VETERAN_ENROLLMENT_SYSTEM_BASE_URI>/ves-associate-gateway-svc/associations/person/1012829228V424035"
+    body:
+      encoding: UTF-8
+      string: '{"associations":[{"address":{"city":"Tampa","country":"USA","line1":"875
+        Updated Blvd","line2":"Unit 532","state":"FL","zipCode":"33726","zipPlus4":"3942"},"name":{"givenName":"UPDATEDFIRSTNOKB","middleName":"UPDATEDMIDDLENOKB","familyName":"UPDATEDLASTNOKB"},"alternatePhone":"2658350023","primaryPhone":"1238835546","relationType":"CHILDINLAW","role":"OTHER_NEXT_OF_KIN","lastUpdateDate":"2025-06-05T20:59:59Z"},{"address":{"city":"Wichita","country":"USA","line1":"875
+        West Blvd","line2":"Apt 3","line3":"Unit 6","state":"KS","zipCode":"67203","zipPlus4":"1234"},"name":{"givenName":"UPDATEDFIRSTECB","middleName":"UPDATEDMIDDLEECB","familyName":"UPDATEDLASTECB"},"alternatePhone":"9563001117","primaryPhone":"6437432434","relationType":"GRANDCHILD","role":"OTHER_EMERGENCY_CONTACT","lastUpdateDate":"2025-06-05T20:59:59Z"},{"address":{"city":"guanajuato","country":"MEX","line1":"NE
+        54th St","line2":"Apt 7","line3":"Unit 1222","provinceCode":"guanajuato","postalCode":"84754"},"name":{"givenName":"UPDATEDFIRSTNOKA","middleName":"UPDATEDMIDDLENOKA","familyName":"UPDATEDLASTNOKA","suffix":"JR."},"alternatePhone":"6544551234","primaryPhone":"4449131234","relationType":"NIECE_NEPHEW","role":"PRIMARY_NEXT_OF_KIN","lastUpdateDate":"2025-06-05T20:59:59Z"},{"address":{"city":"Los
+        Angeles","country":"USA","line1":"28 Parker St","state":"CA","zipCode":"90038","zipPlus4":"1234"},"name":{"givenName":"UPDATEDFIRSTECA","middleName":"UPDATEDMIDDLEECA","familyName":"UPDATEDLASTECA"},"alternatePhone":"1123321232","primaryPhone":"7563627422","relationType":"EXTENDED_FAMILY_MEMBER","role":"EMERGENCY_CONTACT","lastUpdateDate":"2025-06-05T20:59:59Z"}]}'
+    headers:
+      Accept:
+      - application/json
+      Content-Type:
+      - application/json
+      User-Agent:
+      - Vets.gov Agent
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Thu, 05 Jun 2025 21:00:00 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Server:
+      - envoy
+      X-Frame-Options:
+      - SAMEORIGIN
+      Content-Security-Policy:
+      - 'default-src https: data: ''unsafe-inline'' ''unsafe-eval'''
+      X-Envoy-Upstream-Service-Time:
+      - '317'
+      Ssl-Env:
+      - 'On'
+      X-Xss-Protection:
+      - 1; mode=block
+      X-Content-Type-Options:
+      - nosniff
+      Server-Timing:
+      - dtSInfo;desc="0", dtRpid;desc="507197184"
+      Strict-Transport-Security:
+      - max-age=16000000; includeSubDomains; preload;
+    body:
+      encoding: UTF-8
+      string: '{"data":{"associations":[{"role":"OTHER_NEXT_OF_KIN","name":{"givenName":"UPDATEDFIRSTNOKB","middleName":"UPDATEDMIDDLENOKB","familyName":"UPDATEDLASTNOKB"},"status":"UPDATED"},{"role":"OTHER_EMERGENCY_CONTACT","name":{"givenName":"UPDATEDFIRSTECB","middleName":"UPDATEDMIDDLEECB","familyName":"UPDATEDLASTECB"},"status":"UPDATED"},{"role":"PRIMARY_NEXT_OF_KIN","name":{"givenName":"UPDATEDFIRSTNOKA","middleName":"UPDATEDMIDDLENOKA","familyName":"UPDATEDLASTNOKA","suffix":"JR."},"status":"UPDATED"},{"role":"EMERGENCY_CONTACT","name":{"givenName":"UPDATEDFIRSTECA","middleName":"UPDATEDMIDDLEECA","familyName":"UPDATEDLASTECA"},"status":"UPDATED"}]},"messages":[{"description":"The
+        request was successfully processed.","code":"completed_success"}]}'
+  recorded_at: Thu, 05 Jun 2025 21:00:00 GMT
+recorded_with: VCR 6.3.1
